video-service/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-dotenv==1.0.1
httpx==0.28.1
anthropic>=0.49.0
av>=13.0.0
Pillow>=10.0.0
